Remove BladeExample | Add CreateAccountExample
