+++ sdk/examples/CreateAccountExample.cc
@@ -1,0 +1,70 @@
+/*-
+ *
+ * Hedera C++ SDK
+ *
+ * Copyright (C) 2020 - 2022 Hedera Hashgraph, LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+#include "AccountBalance.h"
+#include "AccountBalanceQuery.h"
+#include "AccountCreateTransaction.h"
+#include "Client.h"
+#include "ED25519PrivateKey.h"
+#include "PublicKey.h"
+#include "TransactionRecord.h"
+#include "TransactionResponse.h"
+#include "TransferTransaction.h"
+
+#include <iostream>
+
+using namespace Hedera;
+
+int main(int argc, char** argv)
+{
+  if (argc < 3)
+  {
+    std::cout << "Please input account ID and private key" << std::endl;
+    return 1;
+  }
+
+  // Get a client for the Hedera testnet, and set the operator account ID and key such that all generated transactions
+  // will be paid for by this account and be signed by this key.
+  Client client = Client::forTestnet();
+  std::unique_ptr<PrivateKey> operatorPrivateKey = ED25519PrivateKey::fromString(argv[2]);
+  client.setOperator(AccountId(argv[1]), operatorPrivateKey);
+
+  // Generate a ED25519 private, public key pair
+  const std::unique_ptr<PrivateKey> privateKey = ED25519PrivateKey::generatePrivateKey();
+  const std::shared_ptr<PublicKey> publicKey = privateKey->getPublicKey();
+
+  std::cout << "Generated private key: " << privateKey->toString() << std::endl;
+  std::cout << "Generated public key: " << publicKey->toString() << std::endl;
+
+  // Create a new account with an initial balance of 1000 tinybars. The only required field here is the key.
+  TransactionResponse txResp =
+    AccountCreateTransaction().setKey(publicKey).setInitialBalance(Hbar(1000ULL, HbarUnit::TINYBAR())).execute(client);
+
+  // Get the receipt when it becomes available
+  TransactionReceipt txReceipt;
+  while (!txReceipt.getAccountId().has_value())
+  {
+    txReceipt = txResp.getReceipt(client);
+  }
+
+  const AccountId newAccountId = txReceipt.getAccountId().value();
+  std::cout << "Created new account with ID " << newAccountId.toString() << std::endl;
+
+  return 0;
+}
